Move stress tiers into shared/stress_graphs.rs and include from callers.
